pyproject.toml
[project]
name = "mcp-client-all"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiolimiter>=1.2.1",
    "anyio>=4.12.1",
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.44",
    "cachetools>=7.0.0",
    "fastapi>=0.128.5",
    "fastmcp>=2.14.5",
    "hdbcli>=2.27.23",
    "httpx>=0.28.1",
    "langchain>=1.2.9",
    "langchain-mcp-adapters>=0.2.1",
    "playwright>=1.58.0",
    "prometheus-client>=0.24.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.10",
    "sap-ai-sdk-gen>=6.1.2",
    "solace-pubsubplus>=1.11.0",
    "structlog>=25.5.0",
    "tabulate>=0.9.0",
    "uvicorn[standard]>=0.40.0",
    "yaspin>=3.4.0",
]
